move: constants 파일을 utils->src 디렉토리로 이동
